refactoring project files and update version
