+++ csm_test_utils/autoscaling/__init__.py
@@ -1,0 +1,2 @@
+from .loadbalancer import main as as_lb_main
+from .smn_webhook import main as as_wh_main
